Refactor: push MemoryStore into the app #2714
