+++ data-serving/reusable-data-service/data_service/controller/schema_controller.py
@@ -20,7 +20,9 @@ class SchemaController:
         if name in [f.name for f in existing_fields]:
             raise ConflictError(f"field {name} already exists")
         if type not in Field.acceptable_types:
-            raise PreconditionUnsatisfiedError(f"cannot use {type} as the type of a field")
+            raise PreconditionUnsatisfiedError(
+                f"cannot use {type} as the type of a field"
+            )
         fields_list = [(f.name, f.type, f) for f in existing_fields]
         fields_list.append((name, type, dataclasses.field(init=False, default=None)))
         # re-invent the Case class
